Changed vector type to 16-bit integer signed

+++ Assign.h
@@ -11,25 +11,26 @@
 
 #include <avr/io.h>
 
-typedef int8_t vect_t;
+typedef int16_t vect_t;
 
-#define TMR_OFF 0b000
-#define PSC_1     0b001
-#define PSC_0_8     0b010
-#define PSC_0_64    0b011
-#define PSC_0_256   0b100
-#define PSC_0_1024  0b101
-#define EXT_0_FALLING     0b110
-#define EXT_0_RISING      0b111
-#define PSC_2_8     0b010
-#define PSC_2_32    0b011
-#define PSC_2_64   0b100
-#define PSC_2_128  0b101
-#define PSC_2_256     0b110
+// Timer Prescaler
+#define TMR_OFF         0b000
+#define PSC_1           0b001
+#define PSC_0_8         0b010
+#define PSC_0_64        0b011
+#define PSC_0_256       0b100
+#define PSC_0_1024      0b101
+#define EXT_0_FALLING   0b110
+#define EXT_0_RISING    0b111
+#define PSC_2_8         0b010
+#define PSC_2_32        0b011
+#define PSC_2_64        0b100
+#define PSC_2_128       0b101
+#define PSC_2_256       0b110
 #define PSC_2_1024      0b111
+
 // Motor (ESC) Control Pins/Ports
     // Needed Software PWM !
-
 #define ESC_port PORTD
 #define ESC_dir DDRD
 #define FL_pin PIND3
